wardrobe-3

+++ Task_1/users/users.js
@@ -22,18 +22,23 @@ function logOut() {
 const title = document.querySelector("#main-title");
 const wardrobe = document.querySelector("#wardrobe-table-body");
 
-title.innerHTML = `WARDROBE OF ${usersLogin}`;
+title.innerHTML = `WARDROBE OF<br>${usersLogin}`;
 
-let result;
+let result = [];
 
-db.collection(usersLogin)
-  .get()
-  .then((res) => {
-    result = res;
-    res.docs.forEach((doc) => {
-      renderItems(doc);
-    });
-  });
+// db.collection(usersLogin).onSnapshot((querySnapshot) => {
+//   console.log("here");
+// });
+
+// db.collection(usersLogin)
+//   .get()
+//   .then((res) => {
+//     result = res;
+//     console.log({ res });
+//     res.docs.forEach((doc) => {
+//       renderItems(doc);
+//     });
+//   });
 
 function renderItems(item) {
   wardrobe.innerHTML += `<tr class="user__element" id="${item.id}">
@@ -42,18 +47,52 @@ function renderItems(item) {
   <td>${item.data().size}</td>
   <td>${item.data().purpose}</td>
   <td><img src="${item.data().picture}" alt="Error" class="item_img"></td>
-</tr>`;
+  <td class="btns-wrapper">
+  <button class="edit-btns__el common-btn" onclick="editItem('${item.id}')">Edit</button>
+  <button class="edit-btns__el common-btn" onclick="delItem('${item.id}')">Del</button>
+  </td>
+  </tr>`;
 }
+
+// Realtime stuff
+
+db.collection(usersLogin).onSnapshot(snapshot => {
+  let changes = snapshot.docChanges();
+  changes.forEach(change => {
+    if (change.type == 'added') {
+      renderItems(change.doc)
+    } else if (change.type == 'removed') {
+      let item = wardrobe.querySelector('[id=' + change.doc.id + ']')
+      wardrobe.removeChild(item)
+    } else if (change.type == 'modified') {
+      wardrobe.querySelector('[id=' + change.doc.id + ']').innerHTML = `<td class="user__username">${change.doc.data().name}</td>
+      <td>${change.doc.data().color}</td>
+      <td>${change.doc.data().size}</td>
+      <td>${change.doc.data().purpose}</td>
+      <td><img src="${change.doc.data().picture}" alt="Error" class="item_img"></td>
+      <td class="btns-wrapper">
+      <button class="edit-btns__el common-btn" onclick="editItem('${change.doc.id}')">Edit</button>
+      <button class="edit-btns__el common-btn" onclick="delItem('${change.doc.id}')">Del</button>
+      </td>`
+    }
+  })
+})
+
 
 /// buttons functions
 
 // Add item
 
+const bigAddBtn = document.querySelector("#btn-add");
 const editForm = document.querySelector("#edit-form");
+const smallBtnAdd = document.querySelector("#add-btn");
+const smallBtnEdit = document.querySelector("#edit-btn");
 
 function addNewItem() {
   editForm.style.display = "flex";
-  document.querySelector("#edit-btn").style.display = "none";
+  smallBtnAdd.style.display = "inline";
+  smallBtnEdit.style.display = "none";
+  bigAddBtn.style.display = "none";
 }
 
 function sendNewDoc() {
@@ -64,31 +103,26 @@ function sendNewDoc() {
     purpose: editForm.purpose.value,
     picture: editForm.picture.value,
   });
-  pageRefresh();
+  cancelEditing();
 }
 
 // Edit item
 
-function editItem() {
-  let arr = document.querySelectorAll(".user__element");
-  arr.forEach((item) => {
-    item.setAttribute("onClick", `editDoc('${item.id}')`)
-    item.classList.add('ready-to-choose')
-});
-}
-
 let itemEditing;
 
-function editDoc(idOfItem) {
-  itemEditing = result.docs.find((i) => i.id == idOfItem);
+function editItem(idOfItem) {
+  itemEditing = wardrobe.querySelector('[id=' + idOfItem + ']')
+  let arrOfTds = itemEditing.querySelectorAll('td')
 
   editForm.style.display = "flex";
-  document.querySelector("#add-btn").style.display = "none";
-  editForm.name.value = itemEditing.data().name;
-  editForm.color.value = itemEditing.data().color;
-  editForm.size.value = itemEditing.data().size;
-  editForm.purpose.value = itemEditing.data().purpose;
-  editForm.picture.value = itemEditing.data().picture;
+  smallBtnAdd.style.display = "none";
+  smallBtnEdit.style.display = "inline";
+
+  editForm.name.value = arrOfTds[0].innerHTML;
+  editForm.color.value = arrOfTds[1].innerHTML;
+  editForm.size.value = arrOfTds[2].innerHTML;
+  editForm.purpose.value = arrOfTds[3].innerHTML;
+  editForm.picture.value = arrOfTds[4].children[0].src;
 }
 
 function setDoc() {
@@ -99,28 +133,25 @@ function setDoc() {
     purpose: editForm.purpose.value,
     picture: editForm.picture.value,
   });
-  pageRefresh();
+  cancelEditing();
 }
+
 
 // Delete item
 
-function delItem() {
-  let arr = document.querySelectorAll(".user__element");
-  arr.forEach((item) => {
-    item.setAttribute("onClick", `deleteFromDoc('${item.id}')`)
-    item.classList.add('ready-to-choose')
+function delItem(idOfItem) {
+  if (window.confirm("Are you sure?")) {
+    db.collection(usersLogin).doc(idOfItem).delete();
   }
-  );
-}
-
-function deleteFromDoc(idOfItem) {
-  db.collection(usersLogin).doc(idOfItem).delete();
-  pageRefresh();
 }
 
 
-function pageRefresh() {
-  setTimeout(() => {
-    location.reload();
-  }, 500); // ДУЖЕ ВАЖЛИВО ЗАДАВАТИ ЙОГО ВОТ ТАК, прописувати функцію всрередині
+function cancelEditing() {
+  editForm.style.display = "none";
+  bigAddBtn.style.display = "inline";
+  editForm.name.value = '';
+  editForm.color.value = '';
+  editForm.size.value = '';
+  editForm.purpose.value = '';
+  editForm.picture.value = '';
 }
